+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/session/filter/FilterApi.kt
@@ -32,7 +32,8 @@ internal interface FilterApi {
      * @param body   the Json representation of a FilterBody object
      */
     @POST(NetworkConstants.URI_API_PREFIX_PATH_R0 + "user/{userId}/filter")
-    fun uploadFilter(@Path("userId") userId: String, @Body body: Filter): Call<FilterResponse>
+    fun uploadFilter(@Path("userId") userId: String,
+                     @Body body: Filter): Call<FilterResponse>
 
     /**
      * Gets a filter with a given filterId from the homeserver
@@ -42,5 +43,6 @@ internal interface FilterApi {
      * @return Filter
      */
     @GET(NetworkConstants.URI_API_PREFIX_PATH_R0 + "user/{userId}/filter/{filterId}")
-    fun getFilterById(@Path("userId") userId: String, @Path("filterId") filterId: String): Call<Filter>
+    fun getFilterById(@Path("userId") userId: String,
+                      @Path("filterId") filterId: String): Call<Filter>
 }